+++ playwright/tina-playwright/tests/rich-text-editor/Italic-parsing.spec.ts
@@ -1,0 +1,34 @@
+import { test } from "@playwright/test";
+import { checkMarkdownOutput } from "./markdown-parsing-utils";
+
+test.describe("Rich Text Editor Parsing Spec", () => {
+  test.beforeEach(async ({ page }) => {
+    await page.goto("http://localhost:3000/admin/index.html#/~", {
+      waitUntil: "domcontentloaded",
+    });
+
+    page.click('button[data-test="enter-edit-mode"]');
+  });
+
+  test("Italic is correctly rendered from markdown using *", async ({
+    page,
+  }) => {
+    await checkMarkdownOutput(
+      page,
+      "Testing *Italic* rendering correctly in markdown",
+      "Italic",
+      "em"
+    );
+  });
+
+  test("Italic is correctly rendered from markdown using _", async ({
+    page,
+  }) => {
+    await checkMarkdownOutput(
+      page,
+      "Testing _Italic_ rendering correctly in markdown",
+      "Italic",
+      "em"
+    );
+  });
+});
